# business/tecma-digital-platform/followup-3.0/be-followup-v3/openapi/openapi.v1.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Followup 3.0 REST API",
    "version": "1.0.0",
    "description": "API REST per Followup 3.0 (CRM multi-progetto). Gestione clienti, appartamenti, trattative (rent/sell), calendario, workspaces, progetti, audit e report. Autenticazione JWT (login, refresh). In locale il backend è su porta 8080."
  },
  "servers": [
    {
      "url": "http://localhost:8080/v1",
      "description": "Backend locale (dev)"
    },
    {
      "url": "/v1",
      "description": "Stesso host (proxy)"
    }
  ],
...
  "paths": {
...
    "/workspaces/{workspaceId}/connectors/n8n/config": {
      "get": {
        "tags": [
          "connectors",
          "workspaces"
        ],
        "operationId": "getN8nConnectorConfig",
        "summary": "Legge configurazione n8n del workspace",
        "description": "Richiede integrations.read e workspace con modulo Integrazioni abilitato.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "ID workspace",
            "schema": {
              "type": "string"
            },
            "example": "673abc12def4567890123456"
          }
        ],
        "responses": {
          "200": {
            "description": "config null se assente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "config"
                  ],
                  "properties": {
                    "config": {
                      "nullable": true,
                      "type": "object",
                      "description": "Riga connettore n8n",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "description": "ID documento"
                        },
                        "workspaceId": {
                          "type": "string",
                          "description": "ID workspace"
                        },
                        "connectorId": {
                          "type": "string",
                          "description": "Valore fisso n8n"
                        },
                        "config": {
                          "type": "object",
                          "properties": {
                            "baseUrl": {
                              "type": "string",
                              "description": "URL base istanza n8n"
                            },
                            "apiKeyMasked": {
                              "type": "string",
                              "description": "API key mascherata"
                            },
                            "defaultWorkflowId": {
                              "type": "string",
                              "description": "ID workflow predefinito"
                            }
                          }
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "description": "Aggiornamento UTC"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato"
          },
          "403": {
            "description": "Permesso negato o modulo non abilitato"
          },
          "500": {
            "description": "Errore server"
          }
        }
      },
      "post": {
        "tags": [
          "connectors",
          "workspaces"
        ],
        "operationId": "saveN8nConnectorConfig",
        "summary": "Salva configurazione n8n",
        "description": "Richiede integrations.update. Valida URL e salva baseUrl, apiKey, opzionale defaultWorkflowId.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "ID workspace",
            "schema": {
              "type": "string"
            },
            "example": "673abc12def4567890123456"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "baseUrl",
                  "apiKey"
                ],
                "properties": {
                  "baseUrl": {
                    "type": "string",
                    "format": "uri",
                    "description": "URL base n8n (es. https://n8n.example.com)"
                  },
                  "apiKey": {
                    "type": "string",
                    "description": "API key n8n"
                  },
                  "defaultWorkflowId": {
                    "type": "string",
                    "description": "Workflow ID predefinito per trigger"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Config salvata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "config"
                  ],
                  "properties": {
                    "config": {
                      "type": "object",
                      "description": "Riga con valori mascherati"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validazione fallita (URL o campi)"
          },
          "401": {
            "description": "Non autenticato"
          },
          "403": {
            "description": "Permesso negato"
          },
          "500": {
            "description": "Errore server"
          }
        }
      },
      "delete": {
        "tags": [
          "connectors",
          "workspaces"
        ],
        "operationId": "deleteN8nConnectorConfig",
        "summary": "Rimuove configurazione n8n",
        "description": "Richiede integrations.delete.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "ID workspace",
            "schema": {
              "type": "string"
            },
            "example": "673abc12def4567890123456"
          }
        ],
        "responses": {
          "200": {
            "description": "Esito cancellazione",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "deleted"
                  ],
                  "properties": {
                    "deleted": {
                      "type": "boolean",
                      "description": "true se rimosso"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato"
          },
          "403": {
            "description": "Permesso negato"
          },
          "500": {
            "description": "Errore server"
          }
        }
      }
    },
    "/workspaces/{workspaceId}/connectors/n8n/trigger": {
      "post": {
        "tags": [
          "connectors",
          "workspaces"
        ],
        "operationId": "triggerN8nConnectorWorkflow",
        "summary": "Esegue un workflow n8n via API",
        "description": "Richiede integrations.update e n8n configurato. Usa workflowId nel body o defaultWorkflowId in config.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "ID workspace",
            "schema": {
              "type": "string"
            },
            "example": "673abc12def4567890123456"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workflowId": {
                    "type": "string",
                    "description": "ID workflow n8n (override del default)"
                  },
                  "data": {
                    "type": "object",
                    "additionalProperties": true,
                    "description": "Payload inviato al workflow"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Risposta API execute n8n",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "executionId": {
                      "type": "integer",
                      "description": "ID esecuzione"
                    },
                    "waitingForWebhook": {
                      "type": "boolean",
                      "description": "Workflow in attesa webhook"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "n8n non configurato o workflowId mancante"
          },
          "401": {
            "description": "Non autenticato"
          },
          "403": {
            "description": "Permesso negato"
          },
          "500": {
            "description": "Errore server"
          },
          "502": {
            "description": "Errore upstream n8n"
          }
        }
      }
    },
    "/workspaces/{workspaceId}/connectors/whatsapp/config": {
      "get": {
        "tags": [
          "connectors",
          "workspaces"
        ],
        "operationId": "getWhatsappConnectorConfig",
        "summary": "Legge configurazione WhatsApp Twilio",
        "description": "Richiede integrations.read. Credenziali Twilio per invio WhatsApp; auth token solo mascherato.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "ID workspace",
            "schema": {
              "type": "string"
            },
            "example": "673abc12def4567890123456"
          }
        ],
        "responses": {
          "200": {
            "description": "config null se assente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "config"
                  ],
                  "properties": {
                    "config": {
                      "nullable": true,
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "description": "ID documento"
                        },
                        "workspaceId": {
                          "type": "string",
                          "description": "ID workspace"
                        },
                        "connectorId": {
                          "type": "string",
                          "description": "Valore fisso whatsapp"
                        },
                        "config": {
                          "type": "object",
                          "properties": {
                            "accountSid": {
                              "type": "string",
                              "description": "Twilio Account SID"
                            },
                            "authTokenMasked": {
                              "type": "string",
                              "description": "Token mascherato"
                            },
                            "fromNumber": {
                              "type": "string",
                              "description": "Numero mittente WhatsApp"
                            }
                          }
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "description": "Aggiornamento UTC"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato"
          },
          "403": {
            "description": "Permesso negato o modulo non abilitato"
          },
          "500": {
            "description": "Errore server"
          }
        }
      },
      "post": {
        "tags": [
          "connectors",
          "workspaces"
        ],
        "operationId": "saveWhatsappConnectorConfig",
        "summary": "Salva configurazione WhatsApp Twilio",
        "description": "Richiede integrations.update. Richiede anche modulo Twilio abilitato sul workspace (controllo a salvataggio).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "ID workspace",
            "schema": {
              "type": "string"
            },
            "example": "673abc12def4567890123456"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "accountSid",
                  "authToken",
                  "fromNumber"
                ],
                "properties": {
                  "accountSid": {
                    "type": "string",
                    "description": "Twilio Account SID"
                  },
                  "authToken": {
                    "type": "string",
                    "description": "Auth token Twilio"
                  },
                  "fromNumber": {
                    "type": "string",
                    "description": "Numero mittente (E.164 / whatsapp:)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Config salvata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "config"
                  ],
                  "properties": {
                    "config": {
                      "type": "object",
                      "description": "Riga con valori mascherati"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Body non valido"
          },
          "401": {
            "description": "Non autenticato"
          },
          "403": {
            "description": "Permesso negato o Twilio non abilitato"
          },
          "500": {
            "description": "Errore server"
          }
        }
      },
      "delete": {
        "tags": [
          "connectors",
          "workspaces"
        ],
        "operationId": "deleteWhatsappConnectorConfig",
        "summary": "Rimuove configurazione WhatsApp Twilio",
        "description": "Richiede integrations.delete.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "ID workspace",
            "schema": {
              "type": "string"
            },
            "example": "673abc12def4567890123456"
          }
        ],
        "responses": {
          "200": {
            "description": "Esito cancellazione",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "deleted"
                  ],
                  "properties": {
                    "deleted": {
                      "type": "boolean",
                      "description": "true se rimosso"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato"
          },
          "403": {
            "description": "Permesso negato"
          },
          "500": {
            "description": "Errore server"
          }
        }
      }
    },
    "/workspaces/{workspaceId}/connectors/whatsapp/test": {
      "post": {
        "tags": [
          "connectors",
          "workspaces"
        ],
        "operationId": "sendWhatsappConnectorTestMessage",
        "summary": "Invio prova WhatsApp (Twilio)",
        "description": "Solo amministratore (requireAdmin). Verifica config Twilio e prefisso whatsapp:.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "ID workspace",
            "schema": {
              "type": "string"
            },
            "example": "673abc12def4567890123456"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "to"
                ],
                "properties": {
                  "to": {
                    "type": "string",
                    "description": "Destinatario (E.164 o whatsapp:)"
                  },
                  "body": {
                    "type": "string",
                    "maxLength": 1600,
                    "description": "Testo messaggio (opzionale)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Messaggio accodato/inviato",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "description": "Esito positivo"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Body non valido"
          },
          "401": {
            "description": "Non autenticato"
          },
          "403": {
            "description": "Non amministratore"
          },
          "500": {
            "description": "Errore server"
          }
        }
      }
    },
...
    "/connectors/outlook": {
      "delete": {
        "tags": [
          "connectors"
        ],
        "operationId": "deleteOutlookConnectorCredentials",
        "summary": "Revoca credenziali Outlook (utente corrente)",
        "description": "Richiede integrations.update. Opzionale query workspaceId: se presente e workspace entitled a integrazioni, revoca credenziali per quel contesto.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "query",
            "required": false,
            "description": "ID workspace per revoca contestuale (opzionale)",
            "schema": {
              "type": "string"
            },
            "example": "673abc12def4567890123456"
          }
        ],
        "responses": {
          "200": {
            "description": "Esito revoca",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "deleted"
                  ],
                  "properties": {
                    "deleted": {
                      "type": "boolean",
                      "description": "true se eliminate credenziali"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato"
          },
          "403": {
            "description": "Permesso negato"
          },
          "500": {
            "description": "Errore server"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from login or refresh"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Platform consumer API key (scoped by workspace/project)."
      },
      "signatureWebhookAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-signature-webhook-secret",
        "description": "Segreto condiviso per webhook firme (stesso valore di SIGNATURE_WEBHOOK_SECRET). Equivalente: Authorization Bearer <secret>."
      }
    },
...
  }
}